requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.6.1
ebaysdk==2.2.0
et-xmlfile==1.1.0
exchangelib==5.1.0
fake-useragent==1.4.0
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
lxml==5.1.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.1
parse==1.20.1
pycparser==2.21
pyee==11.1.0
Pygments==2.17.2
pyppeteer==2.0.0
pyquery==2.0.0
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-html==0.10.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sspilib==0.1.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.17.0
